Programa dar optimizuota vektoriams ir dekams. Optimizavimas sulygintas.

+++ src/strukturu_metodai.cpp
@@ -21,7 +21,7 @@ void Studentas::apsk_med()
 {
     if (pazymiai.size() == 0)
         return;
-    Container<uint8_t> visi_pazymiai = pazymiai;
+    Container<int> visi_pazymiai = pazymiai;
     visi_pazymiai.push_back(egzo_rezas);
     rikiavimas(visi_pazymiai); // sort(..) surikiuoja visi_pazymiai vektorių did. tvarka
     int visu_pazymiu_sk = visi_pazymiai.size();
